Move controller HttpConnectionPoolConfigurationManagementController and tests to configuration package.
